more linemod clean-up and fixes; starting on a tool for training linemod templates
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4847 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ recognition/src/linemod.cpp
@@ -69,7 +69,7 @@ createAndAddTemplate (const std::vector<pcl::QuantizableModality*> & modalities,
   SparseQuantizedMultiModTemplate linemod_template;
 
   // select N features from every modality (N = 50, hardcoded; CHANGE this to a parameter!!!)
-  const int nr_features_per_modality = 50;
+  const size_t nr_features_per_modality = 50;
   const size_t nr_modalities = modalities.size();
   for (size_t modality_index = 0; modality_index < nr_modalities; ++modality_index)
   {
